os: move probes.h and rename it to probes_priv.h
Private headers should not be in public include directory.
