+++ dix/events.c
@@ -132,6 +132,7 @@ Equipment Corporation.
 #include "os/client_priv.h"
 #include "os/fmt.h"
 #include "os/log_priv.h"
+#include "os/probes_priv.h"
 #include "Xext/panoramiX.h"
 #include "Xext/panoramiXsrv.h"
 #include "xkb/xkbsrv_priv.h"
@@ -146,7 +147,6 @@ Equipment Corporation.
 #include "dixstruct.h"
 #include "globals.h"
 #include "xace.h"
-#include "probes.h"
 #include "exglobals.h"
 #include "extnsionst.h"
 #include "dispatch.h"
