+++ dix/getevents.c
@@ -43,6 +43,7 @@
 #include "dix/input_priv.h"
 #include "mi/mi_priv.h"
 #include "os/bug_priv.h"
+#include "os/probes_priv.h"
 #include "Xext/panoramiX.h"
 #include "Xext/panoramiXsrv.h"
 
@@ -58,14 +59,11 @@
 #include "eventconvert.h"
 #include "inpututils.h"
 #include "windowstr.h"
-
 #include "xkbsrv.h"
-
 #include "exglobals.h"
 #include "exevents.h"
 #include "extnsionst.h"
 #include "listdev.h"            /* for sizing up DeviceClassesChangedEvent */
-#include "probes.h"
 
 /* Number of motion history events to store. */
 #define MOTION_HISTORY_SIZE 256
